+++ src/main/java/it/denzosoft/jprolog/builtin/io/PeekCode.java
@@ -34,13 +34,15 @@ public class PeekCode implements BuiltIn {
                 throw new PrologEvaluationException("Cannot peek from current input stream: " + currentInputAlias);
             }
 
-            // Use PushbackInputStream to peek without consuming
+            // START_CHANGE: ISS-2025-0193 - Wrap and register PushbackInputStream for reuse
             PushbackInputStream pushbackStream;
             if (inputStream instanceof PushbackInputStream) {
                 pushbackStream = (PushbackInputStream) inputStream;
             } else {
                 pushbackStream = new PushbackInputStream(inputStream);
+                StreamManager.registerInputStream(currentInputAlias, pushbackStream);
             }
+            // END_CHANGE: ISS-2025-0193
 
             int charCode = pushbackStream.read();