+++ src/main/java/it/denzosoft/jprolog/builtin/string/StringCodes.java
@@ -40,9 +40,10 @@ public class StringCodes implements BuiltIn {
             String stringValue = ((PrologString) stringTerm).getStringValue();
             List<Term> codes = new ArrayList<>();
             
-            for (char c : stringValue.toCharArray()) {
-                codes.add(new it.denzosoft.jprolog.core.terms.Number((double) (int) c));
-            }
+            // START_CHANGE: ISS-2025-0193 - Use codePoints for correct supplementary Unicode
+            stringValue.codePoints().forEach(cp ->
+                codes.add(new it.denzosoft.jprolog.core.terms.Number((long) cp)));
+            // END_CHANGE: ISS-2025-0193
             
             Term codesList = createList(codes);
             Map<String, Term> newBindings = new HashMap<>(bindings);
@@ -69,13 +70,10 @@ public class StringCodes implements BuiltIn {
                     throw new PrologEvaluationException("string_codes/2: invalid character code: " + codeValue);
                 }
 
-                // START_CHANGE: ISS-2025-0169 - Fix Unicode truncation for codes > 65535
+                // START_CHANGE: ISS-2025-0193 - Support supplementary Unicode codepoints
                 int intCode = (int) codeValue;
-                if (intCode > Character.MAX_VALUE) {
-                    throw new PrologException(ISOErrorTerms.representationError("character_code", "string_codes/2"));
-                }
-                // END_CHANGE: ISS-2025-0169
-                stringBuilder.append((char) intCode);
+                stringBuilder.append(Character.toChars(intCode));
+                // END_CHANGE: ISS-2025-0193
             }
             
             Term stringResult = new PrologString(stringBuilder.toString());